requirements.txt
google-generativeai>=0.7.2,<1.0
streamlit>=1.28.0
